Refactor inner classes into separate classes
