+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/models/events/EventSetPointBoolean.java
@@ -15,7 +15,7 @@ import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
 import java.util.List;
 
-import static com.powsybl.dynawaltz.ParametersSet.ParameterType.BOOL;
+import static com.powsybl.dynawaltz.parameters.ParameterType.BOOL;
 
 /**
  * @author Mathieu BAGUE {@literal <mathieu.bague at rte-france.com>}
