+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/parameters/ParameterType.java
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+package com.powsybl.dynawaltz.parameters;
+
+/**
+ * @author Marcos de Miguel <demiguelm at aia.es>
+ * @author Florian Dupuy <florian.dupuy at rte-france.com>
+ */
+public enum ParameterType {
+    DOUBLE,
+    INT,
+    BOOL,
+    STRING;
+}
