+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/parameters/ParametersSet.java
@@ -4,13 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynawaltz;
+package com.powsybl.dynawaltz.parameters;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.exceptions.UncheckedXmlStreamException;
 import com.powsybl.commons.xml.XmlUtil;
-import com.powsybl.dynawaltz.parameters.Reference;
 import com.powsybl.dynawaltz.xml.ParametersXml;
 import com.powsybl.dynawaltz.xml.XmlStreamWriterFactory;
 
@@ -22,105 +20,12 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
 
-import static com.powsybl.dynawaltz.ParametersSet.ParameterType.DOUBLE;
-import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_PREFIX;
 import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
 
 /**
  * @author Marcos de Miguel <demiguelm at aia.es>
  */
 public final class ParametersSet {
-
-    public void write(Path parametersPath) throws IOException, XMLStreamException {
-        try (Writer writer = Files.newBufferedWriter(parametersPath, StandardCharsets.UTF_8)) {
-            XMLStreamWriter xmlWriter = XmlStreamWriterFactory.newInstance(writer);
-            try {
-                xmlWriter.writeStartDocument(StandardCharsets.UTF_8.toString(), "1.0");
-                xmlWriter.setPrefix("", DYN_URI);
-                xmlWriter.writeStartElement(DYN_URI, "parametersSet");
-                xmlWriter.writeNamespace("", DYN_URI);
-                for (Map.Entry<String, Set> e : parameterSets.entrySet()) {
-                    writeSet(xmlWriter, e.getKey(), e.getValue());
-                }
-                xmlWriter.writeEndElement();
-                xmlWriter.writeEndDocument();
-            } finally {
-                xmlWriter.close();
-            }
-        }
-    }
-
-    private void writeSet(XMLStreamWriter xmlWriter, String setId, Set set) throws XMLStreamException {
-        xmlWriter.writeStartElement(DYN_URI, "set");
-        xmlWriter.writeAttribute("id", setId);
-        for (Parameter par : set.getParameters().values()) {
-            ParametersXml.writeParameter(xmlWriter, par.getType(), par.getName(), par.getValue());
-        }
-        for (Reference par : set.getReferences()) {
-            ParametersXml.writeReference(xmlWriter, par.getType(), par.getName(), par.getOrigData(), par.getOrigName());
-        }
-        xmlWriter.writeEndElement();
-    }
-
-    public enum ParameterType {
-        DOUBLE,
-        INT,
-        BOOL,
-        STRING;
-    }
-
-    public static class Parameter {
-
-        public Parameter(@JsonProperty("name") String name,
-                         @JsonProperty("type") ParameterType type,
-                         @JsonProperty("value") String value) {
-            this.name = Objects.requireNonNull(name);
-            this.type = Objects.requireNonNull(type);
-            this.value = Objects.requireNonNull(value);
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public ParameterType getType() {
-            return type;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        private final String name;
-        private final ParameterType type;
-        private final String value;
-    }
-
-    public static class Set {
-
-        public void addParameter(String name, ParameterType type, String value) {
-            parameters.put(name, new Parameter(name, type, value));
-        }
-
-        public void addReference(String name, ParameterType type, String origData, String origName) {
-            references.add(new Reference(name, type, origData, origName));
-        }
-
-        public Map<String, Parameter> getParameters() {
-            return parameters;
-        }
-
-        public List<Reference> getReferences() {
-            return references;
-        }
-
-        public Parameter getParameter(String name) {
-            return parameters.get(name);
-        }
-        private final Map<String, Parameter> parameters = new LinkedHashMap<>();
-
-        private final List<Reference> references = new ArrayList<>();
-    }
 
     private final Map<String, Set> parameterSets;
 
@@ -242,5 +147,36 @@ public final class ParametersSet {
                 throw new PowsyblException("Unexpected element: " + xmlReader.getLocalName());
             }
         });
+    }
+
+    public void write(Path parametersPath) throws IOException, XMLStreamException {
+        try (Writer writer = Files.newBufferedWriter(parametersPath, StandardCharsets.UTF_8)) {
+            XMLStreamWriter xmlWriter = XmlStreamWriterFactory.newInstance(writer);
+            try {
+                xmlWriter.writeStartDocument(StandardCharsets.UTF_8.toString(), "1.0");
+                xmlWriter.setPrefix("", DYN_URI);
+                xmlWriter.writeStartElement(DYN_URI, "parametersSet");
+                xmlWriter.writeNamespace("", DYN_URI);
+                for (Map.Entry<String, Set> e : parameterSets.entrySet()) {
+                    writeSet(xmlWriter, e.getKey(), e.getValue());
+                }
+                xmlWriter.writeEndElement();
+                xmlWriter.writeEndDocument();
+            } finally {
+                xmlWriter.close();
+            }
+        }
+    }
+
+    private void writeSet(XMLStreamWriter xmlWriter, String setId, Set set) throws XMLStreamException {
+        xmlWriter.writeStartElement(DYN_URI, "set");
+        xmlWriter.writeAttribute("id", setId);
+        for (Parameter par : set.getParameters().values()) {
+            ParametersXml.writeParameter(xmlWriter, par.getType(), par.getName(), par.getValue());
+        }
+        for (Reference par : set.getReferences()) {
+            ParametersXml.writeReference(xmlWriter, par.getType(), par.getName(), par.getOrigData(), par.getOrigName());
+        }
+        xmlWriter.writeEndElement();
     }
 }
